# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Minimarket API",
    "description": "Contrato OpenAPI para la gestion de productos, carritos y recursos principales del minimarket.",
    "contact": {
      "name": "Maamartinezr",
      "url": "https://github.com/Maamartinezr/Backend-II_Sumativa_S7_Grupo5-"
    },
    "license": {
      "name": "Sin licencia definida",
      "url": "https://github.com/Maamartinezr/Backend-II_Sumativa_S7_Grupo5-"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:9090",
      "description": "Entorno local"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/carrito/{id}": {
      "get": {
        "tags": [
          "Carrito"
        ],
        "summary": "Obtener item del carrito por id",
        "description": "Recupera un item del carrito por su identificador. Requiere autenticacion JWT.",
        "operationId": "obtenerCarritoPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador unico del item del carrito",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Item del carrito encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Carrito"
                },
                "example": {
                  "id": 1,
                  "usuarioId": 1,
                  "productoId": 1,
                  "cantidad": 2
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "2026-07-06T04:12:49.271Z",
                  "status": 401,
                  "error": "Unauthorized",
                  "message": "Full authentication is required to access this resource",
                  "path": "/api/productos"
                }
              }
            }
          },
          "403": {
            "description": "Acceso denegado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "2026-07-06T04:12:49.271Z",
                  "status": 403,
                  "error": "Forbidden",
                  "message": "Access Denied",
                  "path": "/api/productos"
                }
              }
            }
          },
          "404": {
            "description": "Item del carrito no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "2026-07-06T04:12:49.271Z",
                  "status": 404,
                  "error": "Not Found",
                  "message": "Carrito no encontrado con id: 99",
                  "path": "/api/carrito/99"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "2026-07-06T04:12:49.271Z",
                  "status": 500,
                  "error": "Internal Server Error",
                  "message": "Ha ocurrido un error inesperado",
                  "path": "/api/productos"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Carrito"
        ],
        "summary": "Actualizar item del carrito",
        "description": "Actualiza un item existente del carrito. Requiere autenticacion JWT.",
        "operationId": "actualizarCarrito",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador unico del item del carrito",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "description": "Datos actualizados del item de carrito",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Carrito"
              },
              "example": {
                "usuarioId": 1,
                "productoId": 1,
                "cantidad": 2
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Item del carrito actualizado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Carrito"
                },
                "example": {
                  "id": 1,
                  "usuarioId": 1,
                  "productoId": 1,
                  "cantidad": 2
                }
              }
            }
          },
          "400": {
            "description": "Solicitud invalida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "2026-07-06T04:12:49.271Z",
                  "status": 400,
                  "error": "Bad Request",
                  "message": "La solicitud contiene datos invalidos",
                  "path": "/api/productos"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "2026-07-06T04:12:49.271Z",
                  "status": 401,
                  "error": "Unauthorized",
                  "message": "Full authentication is required to access this resource",
                  "path": "/api/productos"
                }
              }
            }
          },
          "403": {
            "description": "Acceso denegado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "2026-07-06T04:12:49.271Z",
                  "status": 403,
                  "error": "Forbidden",
                  "message": "Access Denied",
                  "path": "/api/productos"
                }
              }
            }
          },
          "404": {
            "description": "Item, usuario o producto no encontrados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "2026-07-06T04:12:49.271Z",
                  "status": 404,
                  "error": "Not Found",
                  "message": "Carrito no encontrado con id: 99",
                  "path": "/api/carrito/99"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto de datos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "2026-07-06T04:12:49.271Z",
                  "status": 409,
                  "error": "Conflict",
                  "message": "Conflicto de datos: el recurso ya existe o viola restricciones",
                  "path": "/api/productos"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "2026-07-06T04:12:49.271Z",
                  "status": 500,
                  "error": "Internal Server Error",
                  "message": "Ha ocurrido un error inesperado",
                  "path": "/api/productos"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Carrito"
        ],
        "summary": "Eliminar item del carrito",
        "description": "Elimina un item existente del carrito. Requiere autenticacion JWT.",
        "operationId": "eliminarProductoDelCarrito",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador unico del item del carrito",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "204": {
            "description": "Item del carrito eliminado correctamente"
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "2026-07-06T04:12:49.271Z",
                  "status": 401,
                  "error": "Unauthorized",
                  "message": "Full authentication is required to access this resource",
                  "path": "/api/productos"
                }
              }
            }
          },
          "403": {
            "description": "Acceso denegado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "2026-07-06T04:12:49.271Z",
                  "status": 403,
                  "error": "Forbidden",
                  "message": "Access Denied",
                  "path": "/api/productos"
                }
              }
            }
          },
          "404": {
            "description": "Item del carrito no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "2026-07-06T04:12:49.271Z",
                  "status": 404,
                  "error": "Not Found",
                  "message": "Carrito no encontrado con id: 99",
                  "path": "/api/carrito/99"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "2026-07-06T04:12:49.271Z",
                  "status": 500,
                  "error": "Internal Server Error",
                  "message": "Ha ocurrido un error inesperado",
                  "path": "/api/productos"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-07-06T04:12:49.271Z"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "type": "string",
            "example": "La solicitud contiene datos invalidos"
          },
          "path": {
            "type": "string",
            "example": "/api/productos"
          },
          "validationErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
...
      "Carrito": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "usuarioId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "productoId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "cantidad": {
            "type": "integer",
            "format": "int32",
            "example": 2
          }
        },
        "required": [
          "cantidad",
          "productoId",
          "usuarioId"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}